📝 rename table
